requirements.txt
# Core dependencies for API + Database module only

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Progress bars
tqdm==4.66.1

# Date handling
python-dateutil==2.8.2
